requirements.txt
datasets==2.18.0
emoji==2.10.1
fuzzywuzzy==0.18.0
numpy==1.21.5
OpenCC==1.1.6
pandarallel==1.6.5
pandas==2.0.3
PyYAML==6.0.1
scikit_learn==1.4.1.post1
torch==1.12.0a0+git664058f
tqdm==4.66.2
transformers==4.38.2
